speedometer-app/simulator/requirements.txt
paho-mqtt==1.6.1
numpy==1.24.3